feat: add missing tests

+++ src/TrackEasy.Application/Stations/Shared/GeographicalCoordinatesDto.cs
@@ -1,3 +1,3 @@
 namespace TrackEasy.Application.Stations.Shared;
 
-public sealed record GeographicalCoordinatesDto(int Latitude, int Longitude);
+public sealed record GeographicalCoordinatesDto(decimal Latitude, decimal Longitude);
